rpc-v2/archive: Increase number precision to u128

+++ substrate/client/rpc-spec-v2/src/archive/api.rs
@@ -66,7 +66,7 @@ pub trait ArchiveApi<Hash> {
 	///
 	/// This method is unstable and subject to change in the future.
 	#[method(name = "archive_unstable_finalizedHeight")]
-	fn archive_unstable_finalized_height(&self) -> RpcResult<u64>;
+	fn archive_unstable_finalized_height(&self) -> RpcResult<u128>;
 
 	/// Get the hashes of blocks from the given height.
 	///
@@ -77,7 +77,7 @@ pub trait ArchiveApi<Hash> {
 	///
 	/// This method is unstable and subject to change in the future.
 	#[method(name = "archive_unstable_hashByHeight")]
-	fn archive_unstable_hash_by_height(&self, height: u64) -> RpcResult<Vec<String>>;
+	fn archive_unstable_hash_by_height(&self, height: u128) -> RpcResult<Vec<String>>;
 
 	/// Call into the Runtime API at a specified block's state.
 	///
